Fixed bug where under certain circumstances, the select file button would stop responding
onChange event listener for Navbar button would not trigger after uysing Build Tree status bar button
This was because the value of the button was remaining stuck on the previous file, even though
the displayed name of the open file on the button was changing.

Fixed by setting the event target value to null every time a file is selected by the button, forcing
onChange event to happen every time.

Co-authored-by: Charles Gutwirth <charlesgutwirth@gmail.com>
Co-authored-by: Jordan Hisel <hiseljm@gmail.com>
Co-authored-by: Lindsay Baird <lindsay.a.baird@gmail.com>