Created a new function in Sidebar.tsx - updateTreeView
This function is passed down to the Tree and TreeNode components, and allows
components to update the viewData state held in the Sidebar component.

The function is called whenever a tree node is expanded / contracted, and so
keeps the webview viewData state inline with the extension view data state.

Fixes bug where on closing and reopening the tree expandable menu in the sidebar,
the tree view would default to the initial expanded / contracted state provided by the
extension when the webview is instantiated. Now expanded / contracted state is correctly
remembered when the expandable menu section is minimised / expanded.

Co-authored-by: Charles Gutwirth <charlesgutwirth@gmail.com>
Co-authored-by: Jordan Hisel <hiseljm@gmail.com>
Co-authored-by: Lindsay Baird <lindsay.a.baird@gmail.com>